docker/pyproject.toml
[tool.poetry]
name = "plexe-docker"
version = "0.1.0"
description = "Platform for creating and serving ML models using plexe"
authors = [
    "marcellodebernardi <marcello.debernardi@outlook.com>",
    "vaibs-d <vaibhavatlamarts@gmail.com>"
]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.104.0"
uvicorn = "0.23.2"
pydantic = "2.9.2"
redis = "5.0.1"
pymongo = "4.5.0"
python-dotenv = "1.0.0"
plexe = "0.9.2"
httpx = "0.25.0"
tenacity = "9.0.0"
python-multipart = "0.0.6"
pyyaml = "6.0.1"
black = "^24.10.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.2"
black = "24.10.0"
isort = "5.12.0"
mypy = "1.5.1"
flake8 = "6.1.0"
pytest-cov = "4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false